requirements.txt
adapter-transformers==4.0.0
adapters==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
braceexpand==0.1.7
certifi==2022.12.7
charset-normalizer==2.1.1
clip-benchmark==1.6.1
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
filelock==3.13.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.2.3
h5py==3.11.0
huggingface-hub==0.26.2
idna==3.4
imagecodecs==2023.3.16
imageio==2.35.1
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lit==15.0.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.0
numpy==1.24.1
open_clip_torch==2.29.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pillow==10.2.0
propcache==0.2.0
psutil==6.1.0
pycocoevalcap==1.2
pycocotools==2.0.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.13.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.11
tokenizers==0.20.3
torch==2.0.1+cu118
torch-cluster==1.6.3+pt20cu117
torch-geometric==2.6.1
torch-rs==0.0.4
torch-scatter==2.1.2+pt20cu117
torch-sparse==0.6.18+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.67.0
transformers==4.45.2
triton==2.0.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==1.26.13
wcwidth==0.2.13
webdataset==0.2.100
yarl==1.15.2
zipp==3.20.2
